Crate mongodb_schema_describer

The MongoDB Schema describer.

A common place to query all the possible schema data we can from a MongoDB instance.

Structs

CollectionData

All the information we can fetch per collection.

CollectionId

An id of a collection in the schema.

CollectionWalker

A collection/table in the database.

IndexData

All the information we can scrape per index.

IndexField

A field that is part of an index.

IndexId

An id of an index in the schema.

IndexWalker

An index in the database.

MongoSchema

All the possible information we should scrape out from a MongoDB database.

Enums

IndexFieldProperty

Defines the property in the field that is part of an index.

IndexType

The type of an index.

Functions

describe

Describe the contents of the given database. Only bothers about the schema, meaning the collection names and indexes created. Does a bit of magic to the indexes, so if having a full-text index, the field info is sanitized for Prisma’s use cases. We do not return _fts or _ftsx fields, replacing them with the actual fields used to build the full-text index.

version

Get the version.