Function mongodb_schema_describer::describe

pub async fn describe(client: &Client, db_name: &str) -> Result<MongoSchema>

Describe the contents of the given database. Only bothers about the schema, meaning the collection names and indexes created. Does a bit of magic to the indexes, so if having a full-text index, the field info is sanitized for Prisma’s use cases. We do not return _fts or _ftsx fields, replacing them with the actual fields used to build the full-text index.

Be aware these text fields might not come back in the same order used when initialized.